requirements.txt
anyio==4.3.0
asgiref==3.5.2
cachetools==5.3.3
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.1.0
cloudinary==1.30.0
confusable-homoglyphs==3.2.0
cryptography==37.0.4
cssselect==1.2.0
cssutils==2.10.2
dj-database-url==0.5.0
Django==4.1.2
django-appconf==1.0.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-imagekit==5.0.0
django-on-heroku==1.1.2
django-registration==3.3
django-simple-mail==2.3.5
django-tinymce==3.4.0
djangorestframework==3.14.0
djangorestframework-pagination==1.1
dnspython==2.6.1
drf-yasg==1.21.7
email_validator==2.1.1
gunicorn==20.1.0
h11==0.14.0
html2text==2024.2.26
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
lxml==5.2.1
packaging==24.1
pilkit==3.0
pillow==10.3.0
premailer==3.10.0
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.15
python-dateutil==2.9.0.post0
python-decouple==3.5
pytz==2022.6
pyuploadcare==3.1.0
PyYAML==6.0.1
requests==2.28.1
rfc3986==1.5.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0